Add AVL tree code into data-structure repository
Earler assumption of code being incorrect is wrong.
after investigation avl tree found to be correct one.
Update README.md and Makefile.
